告警清理

+++ plugin/openssl_plugin/key/sym_key_generator/src/sym_key_openssl.c
@@ -186,7 +186,7 @@ static HcfResult HcfDesSymmKeySpiCreate(int32_t keyLen, SymKeyImpl *symKey)
 
     EVP_CIPHER_CTX_free(ctx);
     symKey->keyMaterial.data = keyMaterial;
-    symKey->keyMaterial.len = keyLen;
+    symKey->keyMaterial.len = (size_t)keyLen;
     return HCF_SUCCESS;
 }
